feat: completed form logic
