+++ webapp/package.json
@@ -105,6 +105,7 @@
 		"lodash": "^4.17.21",
 		"nanoid": "^5.0.7",
 		"node-forge": "^1.3.1",
+		"object-to-formdata": "^4.5.1",
 		"pocketbase": "^0.21.3",
 		"prismjs": "^1.29.0",
 		"qrcode-generator": "^1.4.4",
